fix(frontend): Fix dependencies issues
