+++ frontend/src/lib/components/flows/flowInfers.ts
@@ -1,0 +1,58 @@
+import { inferArgs, loadSchemaFromPath } from '$lib/infer'
+import { loadSchemaFlow } from '$lib/scripts'
+import type { Schema } from '$lib/common'
+import { emptySchema } from '$lib/utils'
+import type { FlowModule, InputTransform } from '$lib/gen'
+
+export async function loadSchemaFromModule(module: FlowModule): Promise<{
+	input_transforms: Record<string, InputTransform>
+	schema: Schema
+}> {
+	const mod = module.value
+
+	if (mod.type == 'rawscript' || mod.type === 'script' || mod.type === 'flow') {
+		let schema: Schema
+		if (mod.type === 'rawscript') {
+			schema = emptySchema()
+			await inferArgs(mod.language!, mod.content ?? '', schema)
+		} else if (mod.type == 'script' && mod.path && mod.path != '') {
+			schema = await loadSchemaFromPath(mod.path!, mod.hash)
+		} else if (mod.type == 'flow' && mod.path && mod.path != '') {
+			schema = await loadSchemaFlow(mod.path!)
+		} else {
+			return {
+				input_transforms: {},
+				schema: emptySchema()
+			}
+		}
+
+		const keys = Object.keys(schema?.properties ?? {})
+
+		let input_transforms = mod.input_transforms ?? {}
+
+		if (JSON.stringify(keys.sort()) !== JSON.stringify(Object.keys(input_transforms).sort())) {
+			input_transforms = keys.reduce((accu, key) => {
+				let nv =
+					input_transforms[key] ??
+					(module.id == 'failure' && ['message', 'name'].includes(key)
+						? { type: 'javascript', expr: `error.${key}` }
+						: {
+								type: 'static',
+								value: undefined
+						  })
+				accu[key] = nv
+				return accu
+			}, {})
+		}
+
+		return {
+			input_transforms: input_transforms,
+			schema: schema ?? emptySchema()
+		}
+	}
+
+	return {
+		input_transforms: {},
+		schema: emptySchema()
+	}
+}
