+++ frontend/src/lib/components/flows/flowModuleNextId.ts
@@ -1,0 +1,18 @@
+import type { Flow } from '$lib/gen/models/Flow'
+import { dfs } from './dfs'
+import type { FlowState } from './flowState'
+import { charsToNumber, numberToChars } from './idUtils'
+
+// Computes the next available id
+export function nextId(flowState: FlowState, fullFlow: Flow): string {
+	const allIds = dfs(fullFlow.value.modules, (fm) => fm.id)
+	const max = allIds.concat(Object.keys(flowState)).reduce((acc, key) => {
+		if (key === 'failure' || key.includes('branch') || key.includes('loop')) {
+			return acc
+		} else {
+			const num = charsToNumber(key)
+			return Math.max(acc, num + 1)
+		}
+	}, 0)
+	return numberToChars(max)
+}
